Bug Fix & Minor Update
Fixed a bug that the StatusProvider fails to provide valid cursor for
later use.
Fixed a bug that the status displayed in the widget is not the latest
one in the local database.
In class StatusData,
Made the following fields default:
GET_ALL_ORDER_BY
MAX_CREATED_AT_COLUMNS
DB_TEXT_COLUMNS